requirements.txt
aiofiles==24.1.0
bcrypt==4.2.0
boto3==1.34.155
fastapi==0.112.0
langchain-aws==0.1.15
langchain-community==0.2.11
passlib==1.7.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.8.2
pypdf==4.3.1
python-dotenv==1.0.1
python-multipart==0.0.9
regex==2024.7.24
requests==2.32.3
SQLAlchemy==2.0.32
uvicorn==0.30.5
matplotlib==3.9.1.post1
seaborn==0.13.2
pdfminer.six
PyPDF2